filap-api/requirements.txt
flask==3.1.2
sqlalchemy==2.0.43
flask-sqlalchemy==3.1.1
flask-cors==5.0.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-flask==1.3.0
pytest-cov==5.0.0